feat: 인수테스트 기반 통합 테스트 작성 및 인수 테스트용 기능 분리
