+++ README.md
@@ -85,10 +85,10 @@
     - 메뉴
     - 주문
   - [x] 분석한 도메인 간의 관계를 말로 잘 풀어서 요구사항을 정리 해 본다.
-- [ ] 키친포스의 요구사항을 토대로 테스트 코드를 작성
-  - [ ] `@SpringBootTest`로 통합 테스트 코드 작성
-    - [ ] 메뉴 통합테스트
-    - [ ] 주문 통합테스트
+- [x] 키친포스의 요구사항을 토대로 테스트 코드를 작성
+  - [x] `@SpringBootTest`로 통합 테스트 코드 작성
+    - [x] 메뉴 통합테스트
+    - [x] 주문 통합테스트
   - [x] `@ExtendWith(MockitoExtension.class)` 를 이용해 단위 테스트 코드 작성
     - [x] `Product`
       - [x] create 단위 테스트 작성
