+++ src/test/java/kitchenpos/menu/acceptance/MenuAcceptance.java
@@ -1,0 +1,48 @@
+package kitchenpos.menu.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.domain.Menu;
+import kitchenpos.domain.MenuProduct;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class MenuAcceptance {
+    public static ExtractableResponse<Response> 메뉴_생성_요청(String name, BigDecimal price, Long menuGroupId, List<MenuProduct> menuProducts) {
+        Menu menuRequest = new Menu(name, price, menuGroupId, menuProducts);
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(menuRequest)
+                .when().post("/api/menus")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 메뉴_조회_요청() {
+        return RestAssured
+                .given().log().all()
+                .when().get("/api/menus")
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 메뉴_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 메뉴_생성_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+
+    public static void 메뉴_조회됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+}
