+++ src/test/java/kitchenpos/menu/acceptance/MenuGroupAcceptance.java
@@ -1,0 +1,40 @@
+package kitchenpos.menu.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.domain.MenuGroup;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class MenuGroupAcceptance {
+    public static ExtractableResponse<Response> 메뉴그룹_생성_요청(String name) {
+        MenuGroup menuGroupRequest = new MenuGroup(name);
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(menuGroupRequest)
+                .when().post("/api/menu-groups")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 메뉴그룹_조회_요청() {
+        return RestAssured
+                .given().log().all()
+                .when().get("/api/menu-groups")
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 메뉴그룹_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 메뉴그룹_조회됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+}
