+++ src/test/java/kitchenpos/menu/acceptance/MenuGroupAcceptanceTest.java
@@ -1,17 +1,13 @@
 package kitchenpos.menu.acceptance;
 
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static kitchenpos.menu.acceptance.MenuGroupAcceptance.*;
 
 @DisplayName("메뉴그룹 관련 인수 테스트")
 public class MenuGroupAcceptanceTest extends AcceptanceTest {
@@ -43,33 +39,5 @@ public class MenuGroupAcceptanceTest extends AcceptanceTest {
 
         // then
         메뉴그룹_조회됨(response);
-    }
-
-    public static ExtractableResponse<Response> 메뉴그룹_생성_요청(String name) {
-        MenuGroup menuGroupRequest = new MenuGroup(name);
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(menuGroupRequest)
-                .when().post("/api/menu-groups")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 메뉴그룹_조회_요청() {
-        return RestAssured
-                .given().log().all()
-                .when().get("/api/menu-groups")
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 메뉴그룹_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static void 메뉴그룹_조회됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 }
