+++ src/test/java/kitchenpos/menu/acceptance/ProductAcceptanceTest.java
@@ -1,19 +1,15 @@
 package kitchenpos.menu.acceptance;
 
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import java.math.BigDecimal;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static kitchenpos.menu.acceptance.ProductAcceptance.*;
 
 @DisplayName("상품 관련 인수 테스트")
 public class ProductAcceptanceTest extends AcceptanceTest {
@@ -62,35 +58,5 @@ public class ProductAcceptanceTest extends AcceptanceTest {
 
         // then
         상품_목록_조회됨(response);
-    }
-
-    public static ExtractableResponse<Response> 상품_생성_요청(String name, BigDecimal price) {
-        Product productRequest = new Product(name, price);
-        return RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(productRequest)
-                .when().post("/api/products")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 상품_목록_조회() {
-        return RestAssured.given().log().all()
-                .when().get("/api/products")
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 상품_목록_조회됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 상품_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static void 상품_생성_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
     }
 }
