+++ src/test/java/kitchenpos/order/acceptance/OrderAcceptance.java
@@ -1,0 +1,71 @@
+package kitchenpos.order.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.domain.Menu;
+import kitchenpos.domain.Order;
+import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.OrderTable;
+import org.assertj.core.util.Lists;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class OrderAcceptance {
+    public static ExtractableResponse<Response> 주문_상태_수정_요청(Long orderId, String OrderStatus) {
+        Order orderRequest = new Order();
+        orderRequest.setOrderStatus(OrderStatus);
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(orderRequest)
+                .when().put("/api/orders/{orderId}/order-status", orderId)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 주문_생성_요청(OrderTable orderTable, Menu menu) {
+        OrderLineItem orderLineItem = new OrderLineItem();
+        orderLineItem.setMenuId(menu.getId());
+        Order orderRequest = new Order(orderTable.getId(), Lists.newArrayList(orderLineItem));
+
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(orderRequest)
+                .when().post("/api/orders")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 주문_목록_조회_요청() {
+        return RestAssured
+                .given().log().all()
+                .when().get("/api/orders")
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 주문_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 주문_목록_조회됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 주문_상태_수정됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 주문_상태_수정_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+
+    public static void 주문_생성_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+}
