+++ src/test/java/kitchenpos/order/acceptance/TableGroupAcceptance.java
@@ -1,0 +1,46 @@
+package kitchenpos.order.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.TableGroup;
+import org.assertj.core.util.Lists;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class TableGroupAcceptance {
+    public static ExtractableResponse<Response> 단체테이블_해체_요청(Long tableGroupId) {
+        return RestAssured
+                .given().log().all()
+                .when().delete("/api/table-groups/{tableGroupId}", tableGroupId)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 단체테이블_생성_요청(OrderTable mergeSource1, OrderTable mergeSource2) {
+        TableGroup tableGroupRequest = new TableGroup(Lists.newArrayList(mergeSource1, mergeSource2));
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(tableGroupRequest)
+                .when().post("/api/table-groups")
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 단체테이블_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 단체테이블_해체됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static void 단체테이블_생성_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+}
